Fix exception from log formatting in SQLiteViewStore
insertedCount is AtomicInteger not a regular interger so need to call intValue().

+++ src/main/java/com/couchbase/lite/store/SQLiteViewStore.java
@@ -560,7 +560,7 @@ public class SQLiteViewStore implements ViewStore, QueryRowStore {
                 store.getStorageEngine().update("views", updateValues, "view_id=?", whereArgs);
             }
             Log.v(Log.TAG_VIEW, "...Finished re-indexing (%s) to #%d (deleted %d, added %d)",
-                    viewNames(views), dbMaxSequence, deletedCount, insertedCount);
+                    viewNames(views), dbMaxSequence, deletedCount, insertedCount.intValue());
 
             success = true;
             return new Status(Status.OK);
